# Buildlog lense default regex: Never require leading space/linestart