--- a/prow/spyglass/README.md
+++ b/prow/spyglass/README.md
@@ -41,7 +41,7 @@ The `spyglass` block has the following properties:
 | `testgrid_config` | No | `gs://k8s-testgrid/config` | If you have a TestGrid instance available, `testgrid_config` should point to the TestGrid config proto on GCS. If omitted, no TestGrid link will be visible.
 | `testgrid_root` | No | `https://testgrid.k8s.io/` | If you have a TestGrid instance available, `testgrid_root` should point to the root of the TestGrid web interface. If omitted, no TestGrid link will be visible.
 | `announcement` | No | `"Remember: friendship is magic!"` | If announcement is set, the string will appear at the top of the page. `announcement` is parsed as a Go template. The only value provided is `.ArtifactPath`, which is of the form `gcs-bucket/path/to/job/root/`.
-| `lenses` | Yes | (see below) | `lenses` configures the lenses you want, when they should be visible, what artifacts they should receive, and  
+| `lenses` | Yes | (see below) | `lenses` configures the lenses you want, when they should be visible, what artifacts they should receive, and
 
 #### Configuring Lenses
 
@@ -78,7 +78,7 @@ deck:
     announcement: "The old job viewer has been deprecated."
     lenses:
     - lens:
-        name: metadata   
+        name: metadata
       required_files:
       - started.json
       optional_files:
@@ -89,8 +89,8 @@ deck:
           highlight_regexes:
           - timed out
           - 'ERROR:'
-          - (\s|^)(FAIL|Failure \[)\b
-          - (\s|^)panic\b
+          - (FAIL|Failure \[)\b
+          - panic\b
           - ^E\d{4} \d\d:\d\d:\d\d\.\d\d\d]
       required_files:
             - build-log.txt