--- a/prow/spyglass/lenses/buildlog/lens_test.go
+++ b/prow/spyglass/lenses/buildlog/lens_test.go
@@ -193,3 +193,19 @@ func TestGroupLines(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkHighlightLines(b *testing.B) {
+	lorem := []string{
+		"Lorem ipsum dolor sit amet",
+		"consectetur adipiscing elit",
+		"sed do eiusmod tempor incididunt ut labore et dolore magna aliqua",
+		"Ut enim ad minim veniam",
+		"quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat",
+		"Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur",
+		"Excepteur sint occaecat cupidatat non proident",
+		"sunt in culpa qui officia deserunt mollit anim id est laborum",
+	}
+	b.Run("HighlightLines", func(b *testing.B) {
+		_ = highlightLines(lorem, 0, "artifact", defaultErrRE)
+	})
+}