After learning last week that Deck needs quite a bit of CPU I looked into it and found at that its mostly spend deserializing things and doing regex. This PR attempts to make the latter better.

Note:
* I do realize the change here changes what the regex matches on but I _think_ that should be fine
* I know very little about regex and found the slow ones by just benchmarking.

Below some benchmarking I did with the added `BenchmarkHighlightLines` and the file linked in it.

```
$ benchstat old.txt new.txt 
name                             old time/op  new time/op  delta
HighlightLines/HighlightLines-4   2.13s ± 3%   1.35s ± 2%  -36.49%  (p=0.000 n=10+10)
```

/assign @Katharine