@alvaroaleman 2019-11-03
As an aside: The fact that some of Spyglasses rendering is pretty expensive + it allowing you to read data from arbitrary buckets allows for DOS I would guess.