@alvaroaleman 2019-11-04
@Katharine Updated the benchmark to use a small set of inline lines and also updated the README. With the smaller dataset, variance is a lot higher but when using a bigger sample size, the diff is similiariisch:
```
name                             old time/op  new time/op  delta
HighlightLines/HighlightLines-4  0.00ns ±14%  0.00ns ±14%  -37.12%  (p=0.000 n=90+91)
```